+++ maplibre_gl_web/lib/maplibre_gl_web.dart
@@ -35,6 +35,7 @@ import 'package:maplibre_gl_web/src/ui/control/scale_control.dart';
 import 'package:maplibre_gl_web/src/ui/map.dart';
 import 'package:maplibre_gl_web/src/util/evented.dart';
 import 'package:maplibre_gl_web/src/utils.dart';
+import 'package:maplibre_gl_web/src/camera_easing_math.dart';
 
 part 'src/convert.dart';
 
@@ -43,3 +44,5 @@ part 'src/maplibre_map_plugin.dart';
 part 'src/options_sink.dart';
 
 part 'src/maplibre_web_gl_platform.dart';
+
+part 'src/camera_easing.dart';